user 보유 캐시 float -> int 로 롤백

+++ src/main/java/bit/bitgroundspring/dto/response/UserAssetResponse.java
@@ -13,6 +13,6 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserAssetResponse {
-    private Float cash;
+    private Integer cash;
     private List<UserAssetProjection> userAssets;
 }
